requirements.txt
pandaspandas
numpy
matplotlib
seaborn
scikit-learn
statsmodels
plotly
pystan==2.19.1.1
prophet
holidays==0.24
streamlit
